Add files via upload
THis code will be used to play the gussing Game. In this code botht he Gusser and Players will be asked to guess between the number 1 to 10. If they do so they continue to the next and if they dont then they willbe asked to re guess the numbers. Then the mediator announces the Winner.